+++ tests/Options/PreserveLineBreaks.php
@@ -3,6 +3,7 @@
 declare(strict_types=1);
 
 use PHPHtmlParser\Dom;
+use PHPHtmlParser\Options;
 use PHPUnit\Framework\TestCase;
 
 class PreserveLineBreaks extends TestCase
@@ -10,9 +11,8 @@ class PreserveLineBreaks extends TestCase
     public function testPreserveLineBreakTrue()
     {
         $dom = new Dom();
-        $dom->setOptions([
-            'preserveLineBreaks' => true,
-        ]);
+        $dom->setOptions((new Options())->setPreserveLineBreaks(true));
+
         $dom->loadStr('<div class="stream-container ">
 <div class="stream-item js-new-items-bar-container"> </div> <div class="stream">');
 
@@ -22,9 +22,7 @@ class PreserveLineBreaks extends TestCase
     public function testPreserveLineBreakBeforeClosingTag()
     {
         $dom = new Dom();
-        $dom->setOptions([
-            'preserveLineBreaks' => true,
-        ]);
+        $dom->setOptions((new Options())->setPreserveLineBreaks(true));
         $dom->loadStr('<div class="stream-container "
  ><div class="stream-item js-new-items-bar-container"> </div> <div class="stream">');