+++ tests/OptionsTest.php
@@ -2,8 +2,6 @@
 
 declare(strict_types=1);
 
-use PHPHtmlParser\Dom;
-use PHPHtmlParser\Exceptions\UnknownOptionException;
 use PHPHtmlParser\Options;
 use PHPUnit\Framework\TestCase;
 
@@ -13,137 +11,62 @@ class OptionsTest extends TestCase
     {
         $options = new Options();
 
-        $this->assertTrue($options->whitespaceTextNode);
+        $this->assertTrue($options->isWhitespaceTextNode());
     }
 
     public function testSettingOption()
     {
         $options = new Options();
-        $options->setOptions([
-            'strict' => true,
-        ]);
+        $options->setStrict(true);
 
-        $this->assertTrue($options->strict);
-    }
-
-    public function testAddingOption()
-    {
-        $this->expectException(UnknownOptionException::class);
-
-        $options = new Options();
-        $options->setOptions([
-            'test' => true,
-        ]);
+        $this->assertTrue($options->isStrict());
     }
 
     public function testOverwritingOption()
     {
         $options = new Options();
-        $options->setOptions([
-            'strict' => false,
-        ])->setOptions([
-            'strict'             => true,
-            'whitespaceTextNode' => false,
-        ]);
+        $options->setStrict(false);
+        $options2 = new Options();
+        $options2->setStrict(true);
+        $options2->setWhitespaceTextNode(false);
+        $options->setFromOptions($options2);
 
-        $this->assertTrue($options->get('strict'));
-        $this->assertFalse($options->get('whitespaceTextNode'));
-    }
-
-    public function testGettingNoOption()
-    {
-        $options = new Options();
-        $this->assertEquals(null, $options->get('doesnotexist'));
+        $this->assertTrue($options->isStrict());
+        $this->assertFalse($options->isWhitespaceTextNode());
     }
 
     public function testSetters()
     {
         $options = new Options();
 
-        $options->setOptions([
-            'whitespaceTextNode'     => false,
-            'strict'                 => false,
-            'enforceEncoding'        => null,
-            'cleanupInput'           => false,
-            'removeScripts'          => false,
-            'removeStyles'           => false,
-            'preserveLineBreaks'     => false,
-            'removeDoubleSpace'      => false,
-            'removeSmartyScripts'    => false,
-            'htmlSpecialCharsDecode' => false,
-        ]);
-
         $options->setWhitespaceTextNode(true);
-        $this->assertTrue($options->get('whitespaceTextNode'));
+        $this->assertTrue($options->isWhitespaceTextNode());
 
         $options->setStrict(true);
-        $this->assertTrue($options->get('strict'));
+        $this->assertTrue($options->isStrict());
 
         $options->setEnforceEncoding('utf8');
-        $this->assertEquals('utf8', $options->get('enforceEncoding'));
+        $this->assertEquals('utf8', $options->getEnforceEncoding());
 
         $options->setCleanupInput(true);
-        $this->assertTrue($options->get('cleanupInput'));
+        $this->assertTrue($options->isCleanupInput());
 
         $options->setRemoveScripts(true);
-        $this->assertTrue($options->get('removeScripts'));
+        $this->assertTrue($options->isRemoveScripts());
 
         $options->setRemoveStyles(true);
-        $this->assertTrue($options->get('removeStyles'));
+        $this->assertTrue($options->isRemoveStyles());
 
         $options->setPreserveLineBreaks(true);
-        $this->assertTrue($options->get('preserveLineBreaks'));
+        $this->assertTrue($options->isPreserveLineBreaks());
 
         $options->setRemoveDoubleSpace(true);
-        $this->assertTrue($options->get('removeDoubleSpace'));
+        $this->assertTrue($options->isRemoveDoubleSpace());
 
         $options->setRemoveSmartyScripts(true);
-        $this->assertTrue($options->get('removeSmartyScripts'));
+        $this->assertTrue($options->isRemoveSmartyScripts());
 
         $options->setHtmlSpecialCharsDecode(true);
-        $this->assertTrue($options->get('htmlSpecialCharsDecode'));
-
-        // now reset to false
-
-        $options->setWhitespaceTextNode(false);
-        $this->assertFalse($options->get('whitespaceTextNode'));
-
-        $options->setStrict(false);
-        $this->assertFalse($options->get('strict'));
-
-        $options->setEnforceEncoding(null);
-        $this->assertNull($options->get('enforceEncoding'));
-
-        $options->setCleanupInput(false);
-        $this->assertFalse($options->get('cleanupInput'));
-
-        $options->setRemoveScripts(false);
-        $this->assertFalse($options->get('removeScripts'));
-
-        $options->setRemoveStyles(false);
-        $this->assertFalse($options->get('removeStyles'));
-
-        $options->setPreserveLineBreaks(false);
-        $this->assertFalse($options->get('preserveLineBreaks'));
-
-        $options->setRemoveDoubleSpace(false);
-        $this->assertFalse($options->get('removeDoubleSpace'));
-
-        $options->setRemoveSmartyScripts(false);
-        $this->assertFalse($options->get('removeSmartyScripts'));
-
-        $options->setHtmlSpecialCharsDecode(false);
-        $this->assertFalse($options->get('htmlSpecialCharsDecode'));
-    }
-
-    public function testUnknownOptionDom()
-    {
-        $dom = new Dom();
-        $dom->setOptions([
-            'unknown_option' => true,
-        ]);
-
-        $this->expectException(UnknownOptionException::class);
-        $dom->loadStr('<div></div>');
+        $this->assertTrue($options->isHtmlSpecialCharsDecode());
     }
 }